Revert "Merge branch 'develop/back' into 'feat-BE/PlaceAPI'"
This reverts commit aa3bb4f73a62620c4f94701935e36d5a84ddb979, reversing
changes made to fcee822d09836917b4446c5c5fd2b74f670d8a49.

+++ backend/src/main/java/com/ssafy/fullcourse/domain/sharefullcourse/repository/SharedFCTagRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SharedFCTagRepository extends JpaRepository<SharedFCTag, Long> {
+public interface SharedFCTagRepository extends JpaRepository<SharedFCTag, Integer> {
 
 }
